fixed NPE.
git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@3333 71c3de6d-444a-0410-be80-ed276b4c234a

+++ core/src/main/java/hudson/scm/SubversionSCM.java
@@ -809,7 +809,8 @@ public class SubversionSCM extends SCM implements Serializable {
                 req.setAttribute("message",e.getErrorMessage());
                 rsp.forward(Hudson.getInstance(),"error",req);
             } finally {
-                keyFile.delete();
+                if(keyFile!=null)
+                    keyFile.delete();
                 if(item!=null)
                     item.delete();
             }
